+++ src/core/state/weight_matrix.h
@@ -44,7 +44,7 @@ void clear_diagonal(float *arr, int rows, int cols);
 
 /* Sets delays according to spatial organization */
 void set_delays(OutputType output_type, Connection *conn,
-    float* delays, float velocity,
+    int* delays, float velocity,
     float from_spacing, float to_spacing,
     float x_offset, float y_offset);
